Add more typescript.
