+++ package.json
@@ -29,6 +29,7 @@
     "author": "petterigit, Nikug, MaijaHeiskanen, JuhoHeiskanen, Appelsiini1, Apezkin",
     "license": "ISC",
     "devDependencies": {
+        "@types/lodash": "^4.14.173",
         "@types/twit": "^2.2.30",
         "@typescript-eslint/eslint-plugin": "^4.31.1",
         "@typescript-eslint/parser": "^4.31.1",
